+++ notifications/src/main/scala/org/broadinstitute/dsde/workbench/model/Notification.scala
@@ -193,19 +193,6 @@ object Notifications {
     override val description = "Group Access Requested"
   })
 
-  case class SnapshotRequestSubmittedNotification(recipientUserId: WorkbenchUserId,
-                                                  requestName: String,
-                                                  requestId: String,
-                                                  dateSubmitted: String,
-                                                  requestSummary: String
-  ) extends UserNotification
-  val SnapshotRequestSubmittedNotificationType = register(new NotificationType[SnapshotRequestSubmittedNotification] {
-    override val format: RootJsonFormat[SnapshotRequestSubmittedNotification] =
-      jsonFormat5(SnapshotRequestSubmittedNotification.apply)
-    override val description = "Snapshot Request Submitted"
-    override val alwaysOn = true
-  })
-
   case class SnapshotReadyNotification(recipientUserId: WorkbenchUserId,
                                        snapshotExportLink: String,
                                        snapshotName: String,
@@ -215,6 +202,39 @@ object Notifications {
     override val format: RootJsonFormat[SnapshotReadyNotification] =
       jsonFormat4(SnapshotReadyNotification.apply)
     override val description = "Snapshot Ready"
+    override val alwaysOn = true
+  })
+
+  case class TeaspoonsJobSucceededNotification(recipientUserId: WorkbenchUserId,
+                                               pipelineDisplayName: String,
+                                               jobId: String,
+                                               timeSubmitted: String,
+                                               timeCompleted: String,
+                                               quotaConsumedByJob: String,
+                                               quotaRemaining: String,
+                                               userDescription: String
+  ) extends UserNotification
+  val TeaspoonsJobSucceededNotificationType = register(new NotificationType[TeaspoonsJobSucceededNotification] {
+    override val format: RootJsonFormat[TeaspoonsJobSucceededNotification] =
+      jsonFormat8(TeaspoonsJobSucceededNotification.apply)
+    override val description = "Teaspoons Job Succeeded"
+    override val alwaysOn = true
+  })
+
+  case class TeaspoonsJobFailedNotification(recipientUserId: WorkbenchUserId,
+                                            pipelineDisplayName: String,
+                                            jobId: String,
+                                            errorMessage: String,
+                                            timeSubmitted: String,
+                                            timeCompleted: String,
+                                            quotaConsumedByJob: String,
+                                            quotaRemaining: String,
+                                            userDescription: String
+  ) extends UserNotification
+  val TeaspoonsJobFailedNotificationType = register(new NotificationType[TeaspoonsJobFailedNotification] {
+    override val format: RootJsonFormat[TeaspoonsJobFailedNotification] =
+      jsonFormat9(TeaspoonsJobFailedNotification.apply)
+    override val description = "Teaspoons Job Failed"
     override val alwaysOn = true
   })